feat: close  scope — fault tolerance, mTLS, load testing, monitoring page
Cumulative work covering all four  gaps plus an architectural polish pass.

Phase A — MPC fault tolerance:
- twofa/internal/services/twofaService/fault_tolerance_*_test.go: 6 tests / 12
  subtests covering 1-node-down/timeout, slow-node first-2-wins, all-or-nothing
  setup, disable threshold preservation
- backup_codes.go: parallel bcrypt via errgroup (Setup latency 2.5s → 250ms)
- backup_codes_bench_test.go: serial vs parallel benchmarks (b.Loop, Go 1.24+)
- t.Parallel() on independent setup tests
- docs/03 - Security/MPC Fault Tolerance.md

Phase B — mTLS between services:
- scripts/gen-certs.sh: dev PKI generator (CA + 6 leaf certs, TLS 1.3)
- <svc>/internal/bootstrap/tls.go: loadServer/ClientTLSCredentials helpers
- twofa: mpc_transport.go + mpcclient/ adapter package (DDD port impl)
- middleware/client_auth.go: shared_secret defense-in-depth interceptor
- docker-compose: certgen init container (auto-generates PKI on `up`)
- Drops SECURITY(WR-03) deferred comment
- docs/03 - Security/mTLS.md, ADR-011

Phase C — k6 load testing:
- loadtest/k6/{login,setup-2fa,verify-2fa,mixed}.js + lib/{config,auth,totp}.js
- docker-compose.loadtest.yaml + Makefile load-* targets
- loadtest/REPORT.md: results, bottleneck analysis, scaling recommendations
- Optimizations after first run: bcrypt cost 12→10 for backup codes (security
  rationale documented), Gateway ValidateToken cache via Redis (TTL 10s),
  pgxpool 4×runtime.NumCPU() across all services
- Setup p95: 2.87s → 443ms (×6.5); Setup throughput: 2.4 → 5.6 iter/s

Phase D — Frontend monitoring page:
- gateway/internal/monitoring/: PromClient + Collector + handler with whitelist
  PromQL for 6 services, NaN-safe; new /api/v1/admin/monitoring/snapshot
- frontend atoms: metric-card.tsx, service-status-dot.tsx (Liquid Glass)
- frontend widgets: throughput-overview, service-health-grid, mpc-node-status
- app/(protected)/monitoring/page.tsx + use-monitoring hook (10s polling)
- Navbar: Dashboard/Monitoring tabs with active state
- i18n: ru/en monitoring section

Polish:
- .golangci.yml v2 with 33 linters; all 4 services pass with 0 issues
- Makefile golangci-lint target
- IdentityResolver interface (cachedResolver + directResolver) — Auth
  middleware no longer depends on concrete *TokenCache
- Deps.BackupCodeBcryptCost — bcrypt.MinCost in tests drops twofaService
  test wall time from 270s → 2.3s (with race)
- fatalf helpers in auth/twofa/mpc bootstrap (deferred cancel before os.Exit)
- Pre-existing fixes surfaced by linter: unconvert in totp, errcheck on
  Close/Rollback paths
- workspace/ → docs/ (committed, no longer gitignored), wikilinks → relative
  markdown links
- README updates across root + 4 services pointing to new docs/ structure
- Host port exposure trimmed: only frontend:3000, gateway:8080, grafana:3001
- Database wiped at end of session for clean snapshot

+++ twofa/internal/crypto/totp/totp.go
@@ -83,7 +83,7 @@ func validateOTPWithCounterAt(secret []byte, code string, unixTime int64) (bool,
 			return false, 0
 		}
 	}
-	counter := int64(unixTime) / 30
+	counter := unixTime / 30
 	codeBytes := []byte(code)
 
 	matchCurr := subtle.ConstantTimeCompare([]byte(hotp(secret, uint64(counter))), codeBytes)
